refactor: move random number handling to random.c, perlrand.h
Make it easier to find the bits of perl that deal with random number
generation and seeding.
