+++ proto.h
@@ -12864,13 +12864,6 @@ S_mess_alloc(pTHX)
         Perl_attribute_nonnull_aTHX_;
 # define PERL_ARGS_ASSERT_MESS_ALLOC
 
-static U64
-S_splitmix64(pTHX_ U64 *state)
-        Perl_attribute_nonnull_aTHX_
-        Perl_attribute_nonnull_(pTHX_1);
-# define PERL_ARGS_ASSERT_SPLITMIX64            \
-        assert(state)
-
 static SV *
 S_with_queued_errors(pTHX_ SV *ex)
         Perl_attribute_nonnull_aTHX_
