+++ util.h
@@ -62,22 +62,7 @@ Hence, for example, C<ibcmp()> is S<C<(! foldEQ())>>
 #define ibcmp_utf8(s1, pe1, l1, u1, s2, pe2, l2, u2) \
                     cBOOL(! foldEQ_utf8(s1, pe1, l1, u1, s2, pe2, l2, u2))
 
-typedef struct { uint64_t state;  uint64_t inc; } pcg64_random_t;
-
-#define PL_RANDOM_STATE_TYPE pcg64_random_t
-
-#define Perl_pcg64_seed(seed) (Perl_pcg64_seed_r(&PL_random_state, (seed)))
-#define Perl_pcg64_random_NV() \
-    (Perl_pcg64_random_NV_r(&PL_random_state))
-
-#ifdef PERL_CORE
-/* uses a different source of randomness to avoid interfering with the results
- * of rand() */
-#define Perl_internal_randd() \
-    (Perl_pcg64_random_NV_r(&PL_internal_random_state))
-#define Perl_internal_rand_seed(seed) \
-    (Perl_pcg64_seed_r(&PL_internal_random_state, (seed)))
-#endif
+#include "perlrand.h"
 
 #ifdef USE_C_BACKTRACE
