+++ win32/GNUmakefile
@@ -972,6 +972,7 @@ MICROCORE_SRC	=		\
 		..\peep.c	\
 		..\perly.c	\
 		..\pp_sort.c	\
+		..\random.c	\
 		..\reentr.c	\
 		..\run.c	\
 		..\scope.c	\
@@ -1012,6 +1013,7 @@ CORE_NOCFG_H	=		\
 		..\opcode.h	\
 		..\perl.h	\
 		..\perlapi.h	\
+		..\perlrand.h	\
 		..\perlsdio.h	\
 		..\perly.h	\
 		..\pp.h		\
